+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[0.3.14]
+### Added
+- Log important things
+### Changes
+- Cancel realtime requests in the foreach if we are stopping
+
 [0.3.12]
 ### Changed
 - Catch OverflowException